+++ pkg/operator/operator_deployment.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -201,7 +201,6 @@ func (o *Operator) makeDeploymentConfigAndDeps() (deployment.Config, deployment.
 		AllowChaos:                o.Config.AllowChaos,
 		ScalingIntegrationEnabled: o.Config.ScalingIntegrationEnabled,
 		ReconciliationDelay:       o.Config.ReconciliationDelay,
-		Scope:                     o.Scope,
 	}
 	deps := deployment.Dependencies{
 		Client:        o.Client,