+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Maintenance) Unify References via golangci-linter
 - (Feature) Add Default Container Mods
 - (Documentation) Improve Charts and Upgrade Documentation
+- (Feature) AutoDiscover Operator Access
 
 ## [1.2.47](https://github.com/arangodb/kube-arangodb/tree/1.2.47) (2025-03-28)
 - (Bugfix) Use Profile Annotations